requirements.txt
django~=3.2.1
requests~=2.24.0
beautifulsoup4~=4.9.3
djangorestframework~=3.12.4
drf-generators~=0.5.0
psycopg2-binary~=2.8.6
django-postgres-copy~=2.6.0
